Make board file support more packaging friendly
The current way board support is done with libsoc doesn't play well with
distribution packaging logic. This adds a new "--with-board-configs"
option that will copy all the board config files in to
$PREFIX/share/libsoc.

If the --enable-board flag is used in conjunction with this flag a
symlink will be created rather than copying the file to etc.

Automake doesn't have any native support for recursively copying the
contents of a directory, so added some custom logic in
install-data-local and uninstall-local to handle this.

Signed-off-by: Andy Doan <andy.doan@linaro.org>