gpio,pwm: add delay to allow udev rules to complete
After exporting a GPIO or PWM, we need to give some time for udev rules
to complete, before actually trying to access the newly created entries
in the sysfs.

This solves a problem, where udev was setting the mode and group of
the newly created files, so they are accessible for users (not root)
belonging to that group.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>